+++ emg3d/io.py
@@ -19,7 +19,6 @@ Utility functions for writing and reading data.
 
 import os
 import json
-import warnings
 from datetime import datetime
 
 import numpy as np
@@ -46,7 +45,7 @@ KNOWN_CLASSES = {
 }
 
 
-def save(fname, backend=None, compression="gzip", **kwargs):
+def save(fname, compression="gzip", **kwargs):
     """Save surveys, meshes, models, fields, and more to disk.
 
     Serialize and save data to disk in different formats (see parameter
@@ -74,9 +73,6 @@ def save(fname, backend=None, compression="gzip", **kwargs):
           file. Default format if `h5py` is not installed.
         - `.json`: Uses `json` to store inputs to a hierarchical, plain text
           file.
-
-    backend : deprecated
-        Set the appropriate file-ending in `fname` instead.
 
     compression : int or str, optional
         Passed through to h5py, default is 'gzip'.
@@ -120,37 +116,18 @@ def save(fname, backend=None, compression="gzip", **kwargs):
     # sorted TensorMesh, Field, and Model instances.
     data = _dict_serialize(kwargs, collect_classes=collect_classes)
 
-    # Deprecated backends.
-    if backend is not None:
-        mesg = ("\n    The use of `backend` is deprecated and will be removed."
-                "\n    Use the file-endings [`.h5`, `.npz`, `.json`] instead.")
-        warnings.warn(mesg, DeprecationWarning)
-
-        if backend == 'numpy':
-            backend = 'npz'
-        elif backend == 'h5py':
-            backend = 'h5'
+    # Get extension from `fname`.
+    if full_path.split('.')[-1] in ['npz', 'h5', 'json']:
+        extension = full_path.split('.')[-1]
+    else:  # Fallback to default.
+        if isinstance(h5py, str):
+            extension = 'npz'
         else:
-            backend = backend
+            extension = 'h5'
+        full_path += '.'+extension
 
-        # Add file-ending if necessary.
-        if not full_path.endswith('.'+backend):
-            full_path += '.'+backend
-
-    else:
-
-        # Get backend from `fname`.
-        if full_path.split('.')[-1] in ['npz', 'h5', 'json']:
-            backend = full_path.split('.')[-1]
-        else:  # Fallback to default.
-            if isinstance(h5py, str):
-                backend = 'npz'
-            else:
-                backend = 'h5'
-            full_path += '.'+backend
-
-    # Save data depending on the backend.
-    if backend == "npz":
+    # Save data depending on the extension.
+    if extension == "npz":
 
         # Convert hierarchical dict to a flat dict.
         data = _dict_flatten(data)
@@ -158,7 +135,7 @@ def save(fname, backend=None, compression="gzip", **kwargs):
         # Store flattened data.
         np.savez_compressed(full_path, **data)
 
-    elif backend == "h5":
+    elif extension == "h5":
 
         # Check if h5py is installed.
         if isinstance(h5py, str):
@@ -168,7 +145,7 @@ def save(fname, backend=None, compression="gzip", **kwargs):
         with h5py.File(full_path, "w") as h5file:
             _hdf5_add_to(data, h5file, compression)
 
-    elif backend == "json":
+    elif extension == "json":
 
         # Move arrays to lists and decompose complex data.
         data = _dict_dearray_decomp(data)
@@ -178,7 +155,7 @@ def save(fname, backend=None, compression="gzip", **kwargs):
             json.dump(data, f, indent=json_indent)
 
     else:
-        raise ValueError(f"Unknown backend '{backend}'.")
+        raise ValueError(f"Unknown extension '{extension}'.")
 
     # Print file info.
     if verb > 0:
@@ -197,8 +174,7 @@ def load(fname, **kwargs):
     Parameters
     ----------
     fname : str
-        File name including extension. Used backend depends on the file
-        extensions:
+        File name including extension. Possibilities:
 
         - '.npz': numpy-binary
         - '.h5': h5py-binary (needs `h5py`)
@@ -696,7 +672,7 @@ def _hdf5_get_from(h5file):
 def _compare_dicts(dict1, dict2, verb=False, **kwargs):
     """Return True if the two dicts `dict1` and `dict2` are the same.
 
-    Private method, not foolproof. Useful for developing new backends.
+    Private method, not foolproof. Useful for developing new extensions.
 
     If `verb=True`, it prints it key starting with the following legend: